Version 1.0.1 SNAPSHOT

+++ Student_management/src/main/java/com/hieupc/student_management/repository/CourseRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CourseRepository extends JpaRepository<Course, Long> {
+public interface CourseRepository extends JpaRepository<Course, Integer> {
 
     @Query("SELECT c FROM Course c WHERE c.name = :name")
     Course queryByName(@Param("name") String name);
